moving common functions to utils.rs
