+++ crates/evm/src/tests/call_tests.rs
@@ -19,12 +19,15 @@ use crate::smart_contracts::{
     SimpleStorageContract, TestContract,
 };
 use crate::tests::test_signer::TestSigner;
-use crate::tests::utils::{config_push_contracts, get_evm};
+use crate::tests::utils::{
+    config_push_contracts, create_contract_message, create_contract_message_with_fee,
+    create_contract_transaction, get_evm, get_evm_config, get_evm_config_starting_base_fee,
+    publish_event_message, set_arg_message,
+};
 use crate::tests::DEFAULT_CHAIN_ID;
 use crate::{
     AccountData, EvmConfig, RlpEvmTransaction, BASE_FEE_VAULT, L1_FEE_VAULT, PRIORITY_FEE_VAULT,
 };
-
 type C = DefaultContext;
 
 #[test]
@@ -876,33 +879,6 @@ fn test_block_gas_limit() {
     );
 }
 
-pub fn create_contract_message<T: TestContract>(
-    dev_signer: &TestSigner,
-    nonce: u64,
-    contract: T,
-) -> RlpEvmTransaction {
-    dev_signer
-        .sign_default_transaction(TxKind::Create, contract.byte_code(), nonce, 0)
-        .unwrap()
-}
-
-pub(crate) fn create_contract_message_with_fee<T: TestContract>(
-    dev_signer: &TestSigner,
-    nonce: u64,
-    contract: T,
-    max_fee_per_gas: u128,
-) -> RlpEvmTransaction {
-    dev_signer
-        .sign_default_transaction_with_fee(
-            TxKind::Create,
-            contract.byte_code(),
-            nonce,
-            0,
-            max_fee_per_gas,
-        )
-        .unwrap()
-}
-
 pub(crate) fn create_contract_message_with_priority_fee<T: TestContract>(
     dev_signer: &TestSigner,
     nonce: u64,
@@ -922,35 +898,7 @@ pub(crate) fn create_contract_message_with_priority_fee<T: TestContract>(
         .unwrap()
 }
 
-pub(crate) fn create_contract_transaction<T: TestContract>(
-    dev_signer: &TestSigner,
-    nonce: u64,
-    contract: T,
-) -> RlpEvmTransaction {
-    dev_signer
-        .sign_default_transaction(TxKind::Create, contract.byte_code(), nonce, 0)
-        .unwrap()
-}
-
 fn set_selfdestruct_arg_message(
-    contract_addr: Address,
-    dev_signer: &TestSigner,
-    nonce: u64,
-    set_arg: u32,
-) -> RlpEvmTransaction {
-    let contract = SimpleStorageContract::default();
-
-    dev_signer
-        .sign_default_transaction(
-            TxKind::Call(contract_addr),
-            contract.set_call_data(set_arg),
-            nonce,
-            0,
-        )
-        .unwrap()
-}
-
-pub(crate) fn set_arg_message(
     contract_addr: Address,
     dev_signer: &TestSigner,
     nonce: u64,
@@ -1013,75 +961,6 @@ fn selfdestruct_message(
             0,
         )
         .unwrap()
-}
-
-pub(crate) fn publish_event_message(
-    contract_addr: Address,
-    signer: &TestSigner,
-    nonce: u64,
-    message: String,
-) -> RlpEvmTransaction {
-    let contract = LogsContract::default();
-
-    signer
-        .sign_default_transaction(
-            TxKind::Call(contract_addr),
-            contract.publish_event(message),
-            nonce,
-            0,
-        )
-        .unwrap()
-}
-
-pub(crate) fn get_evm_config(
-    signer_balance: U256,
-    block_gas_limit: Option<u64>,
-) -> (EvmConfig, TestSigner, Address) {
-    let dev_signer: TestSigner = TestSigner::new_random();
-
-    let contract_addr = address!("819c5497b157177315e1204f52e588b393771719");
-    let mut config = EvmConfig {
-        data: vec![AccountData {
-            address: dev_signer.address(),
-            balance: signer_balance,
-            code_hash: KECCAK_EMPTY,
-            code: Bytes::default(),
-            nonce: 0,
-            storage: Default::default(),
-        }],
-        spec: vec![(0, SpecId::SHANGHAI)].into_iter().collect(),
-        block_gas_limit: block_gas_limit.unwrap_or(ETHEREUM_BLOCK_GAS_LIMIT),
-        ..Default::default()
-    };
-    config_push_contracts(&mut config, None);
-    (config, dev_signer, contract_addr)
-}
-
-pub(crate) fn get_evm_config_starting_base_fee(
-    signer_balance: U256,
-    block_gas_limit: Option<u64>,
-    starting_base_fee: u64,
-) -> (EvmConfig, TestSigner, Address) {
-    let dev_signer: TestSigner = TestSigner::new_random();
-
-    let contract_addr = address!("819c5497b157177315e1204f52e588b393771719");
-    let mut config = EvmConfig {
-        data: vec![AccountData {
-            address: dev_signer.address(),
-            balance: signer_balance,
-            code_hash: KECCAK_EMPTY,
-            code: Bytes::default(),
-            nonce: 0,
-            storage: Default::default(),
-        }],
-        spec: vec![(0, SpecId::SHANGHAI)].into_iter().collect(),
-        block_gas_limit: block_gas_limit.unwrap_or(ETHEREUM_BLOCK_GAS_LIMIT),
-        starting_base_fee,
-        coinbase: PRIORITY_FEE_VAULT,
-        ..Default::default()
-    };
-    config_push_contracts(&mut config, None);
-    (config, dev_signer, contract_addr)
 }
 
 #[test]
